cleaned Up api/ change some route names
